Adds LocalStorage for data, monthly contribution and date
I implemented localStorage for the data so when the user returns there information is saved. I also had the localStorage keep the monthly contribution as
I am not saving that in the data object itself and without saving and retrieving the numbers don't add up even though the information is stored there.
Lastly I stored the date as well for localStorage in case it is needed.