Adjusts Data Date Just in Adjust Start Month Option
To start the process of being able to adjust the months I needed to first adjust the arrangement
of the months. I accomplished that with Adjust Start Month. I will need to do setData of that
new Array of objects but I need to take into account the initial funding applying to the start
month but most importantly the data selections for expense and contributions as right now
that was assuming January was the start date. Also I need to look at my handleUpdateDataState
to confirm how I am handling that update if changing the months will affect that.